+++ src/main/java/com/cloudstore/server/.settings/org.eclipse.jdt.core.prefs
@@ -1,0 +1,5 @@
+#Sat Apr 04 12:28:27 CEST 2026
+eclipse.preferences.version=1
+org.eclipse.jdt.core.compiler.codegen.targetPlatform=17
+org.eclipse.jdt.core.compiler.compliance=17
+org.eclipse.jdt.core.compiler.source=17
