+++ src/main/java/com/cloudstore/server/dao/impl/TransactionDAOImpl.java
@@ -1,9 +1,9 @@
-package com.cloudstore.server.dao.impl;
+package dao.impl;
 
-import com.cloudstore.server.dao.interfaces.TransactionDAO;
-import com.cloudstore.server.model.entities.Transaction;
-import com.cloudstore.server.model.entities.Product;
-import com.cloudstore.server.utils.DatabaseConnection;
+import dao.interfaces.TransactionDAO;
+import model.entities.Transaction;
+import model.entities.Product;
+import utils.DatabaseConnection;
 
 import java.sql.*;
 import java.time.LocalDateTime;
