+++ src/main/java/com/cloudstore/server/dao/impl/UserDAOImpl.java
@@ -1,9 +1,9 @@
-package com.cloudstore.server.dao.impl;
+package dao.impl;
 
-import com.cloudstore.server.dao.interfaces.UserDAO;
-import com.cloudstore.server.model.entities.User;
-import com.cloudstore.server.model.entities.Permission;
-import com.cloudstore.server.utils.DatabaseConnection;
+import dao.interfaces.UserDAO;
+import model.entities.User;
+import model.entities.Permission;
+import utils.DatabaseConnection;
 
 import java.sql.*;
 import java.util.ArrayList;
