+++ src/main/java/com/cloudstore/server/dao/interfaces/UserDAO.java
@@ -1,6 +1,6 @@
-package com.cloudstore.server.dao.interfaces;
+package dao.interfaces;
 
-import com.cloudstore.server.model.entities.User;
+import model.entities.User;
 import java.sql.SQLException;
 import java.util.List;
 import java.util.Optional;
