+++ src/main/java/com/cloudstore/server/model/entities/Product.java
@@ -1,4 +1,4 @@
-package com.cloudstore.server.model.entities;
+package model.entities;
 
 public record Product (
     int id,
