+++ src/main/java/com/cloudstore/server/model/entities/Transaction.java
@@ -1,4 +1,4 @@
-package com.cloudstore.server.model.entities;
+package model.entities;
 
 import java.time.LocalDateTime;
