+++ src/main/java/com/cloudstore/server/model/entities/User.java
@@ -1,4 +1,4 @@
-package com.cloudstore.server.model.entities;
+package model.entities;
 
 public record User (
    String nickname,
