+++ src/main/java/com/cloudstore/server/service/auth/exception/JwtExpiredException.java
@@ -1,4 +1,4 @@
-package com.cloudstore.server.service.auth.exception;
+package service.auth.exception;
 
 public class JwtExpiredException extends RuntimeException {
