+++ src/main/java/com/cloudstore/server/service/exception/ServiceException.java
@@ -1,4 +1,4 @@
-package com.cloudstore.server.service.exception;
+package service.exception;
 
 public class ServiceException extends Exception {
