+++ src/main/java/com/cloudstore/server/service/impl/TransactionServiceImpl.java
@@ -1,12 +1,12 @@
-package com.cloudstore.server.service.impl;
+package service.impl;
 
-import com.cloudstore.server.dao.impl.TransactionDAOImpl;
-import com.cloudstore.server.dao.interfaces.TransactionDAO;
-import com.cloudstore.server.model.dto.TransactionDTO;
-import com.cloudstore.server.model.entities.Transaction;
-import com.cloudstore.server.service.exception.ServiceException;
-import com.cloudstore.server.service.interfaces.TransactionService;
-import com.cloudstore.server.service.mapper.DTOMapper;
+import dao.impl.TransactionDAOImpl;
+import dao.interfaces.TransactionDAO;
+import model.dto.TransactionDTO;
+import model.entities.Transaction;
+import service.exception.ServiceException;
+import service.interfaces.TransactionService;
+import service.mapper.DTOMapper;
 
 import java.sql.SQLException;
 import java.time.LocalDateTime;
