+++ src/main/java/com/cloudstore/server/service/interfaces/CartService.java
@@ -1,7 +1,7 @@
-package com.cloudstore.server.service.interfaces;
+package service.interfaces;
 
-import com.cloudstore.server.model.dto.TransactionDTO;
-import com.cloudstore.server.service.exception.ServiceException;
+import model.dto.TransactionDTO;
+import service.exception.ServiceException;
 
 import java.util.Map;
