+++ src/main/java/com/cloudstore/server/service/interfaces/UserService.java
@@ -1,7 +1,7 @@
-package com.cloudstore.server.service.interfaces;
+package service.interfaces;
 
-import com.cloudstore.server.model.dto.UserDTO;
-import com.cloudstore.server.service.exception.ServiceException;
+import model.dto.UserDTO;
+import service.exception.ServiceException;
 
 import java.util.List;
 import java.util.Optional;
